Move client code to client class
